Split out tests too
