+++ logging/src/test/scala/com/ovoenergy/effect/LoggingTest.scala
@@ -1,71 +1,26 @@
 package com.ovoenergy.effect
 
 import cats.Applicative
-import cats.data.{Writer, WriterT}
+import cats.data.WriterT
 import cats.effect.IO
 import cats.instances.list._
 import cats.syntax.flatMap._
-import com.ovoenergy.effect.Logging.{Debug, Error, Info, Log, _}
+import com.ovoenergy.effect.Logging._
 import org.scalatest.{Inspectors, Matchers, WordSpec}
-import cats.syntax.applicative._
-import cats.instances.option._
-import Logging._
 
 class LoggingTest extends WordSpec with Matchers with Inspectors {
 
   type LogWriter[A] = WriterT[IO, List[Log], A]
-  type MdcWriter[A] = WriterT[IO, List[(String, String)], A]
-  type TracedLogger[A] = Traced[LogWriter, A]
-  type TracedMdc[A] = Traced[MdcWriter, A]
-  val token = TraceToken("foo")
-
-  implicit val log: Logging[TracedLogger] =
-    tracedInstance[LogWriter](token.pure[LogWriter], (log, _) => WriterT.tell(List(log)))
-
-  def run[A](log: TracedLogger[A]): Writer[List[Log], A] = {
-    val (l, a) = log.runEmptyA.run.unsafeRunSync
-    Writer(l, a)
-  }
-
-  "Tracing" should {
-
-    "Store MDC info" in {
-      run(putMdc("foo" -> "bar") >> mdc).value.get("foo") shouldBe Some("bar")
-    }
-
-    "Create a trace token if one is not set then use the same one thereafter" in {
-        val result = run(
-          for {
-            token1 <- Logging.token
-            token2 <- Logging.token
-          } yield List(token1, token2)
-        ).value
-
-      result.head.value.nonEmpty shouldBe true
-      result.head shouldBe result.tail.head
-    }
-
-    "Include MDC info as a parameter to the log function" in {
-      implicit val log: Logging[TracedMdc] = tracedInstance(token.pure[MdcWriter], (_, m) => WriterT.tell(m.toList))
-      val result = (putMdc[TracedMdc]("foo" -> "bar") >> log.log(Info("blah"))).runEmptyA.written.unsafeRunSync
-      result shouldBe List("foo" -> "bar", "traceToken" -> token.value)
-    }
-
-    "Overwrite any clashing MDC / logging tags" in {
-      implicit val log: Logging[TracedMdc] = tracedInstance(token.pure[MdcWriter], (_, m) => WriterT.tell(m.toList))
-      val result = putMdc[TracedMdc]("foo" -> "bar") >> log.log(Info("blah"), Map("foo" -> "baz"))
-      result.runEmptyA.written.unsafeRunSync shouldBe List("foo" -> "baz", "traceToken" -> token.value)
-    }
-  }
+  implicit val log: Logging[LogWriter] = (l, _) => WriterT.tell(List(l))
 
   "Logging syntax" should {
 
     "log before running the actual effect" in {
       val message = Debug("Testing logging")
       val expected = List(message, Info("boo"))
-      val toLog = Applicative[TracedLogger].unit
+      val toLog = Applicative[LogWriter].unit
       val events = (toLog >> log.log(Info("boo"))).log(message)
-      events.runEmptyA.written.unsafeRunSync shouldEqual expected
+      events.written.unsafeRunSync shouldEqual expected
     }
 
     "log exceptions" in {
@@ -75,9 +30,9 @@ class LoggingTest extends WordSpec with Matchers with Inspectors {
         Error("another", new RuntimeException("ex2"))
       )
 
-      val toLog = Applicative[TracedLogger].unit
+      val toLog = Applicative[LogWriter].unit
       val events = errors.foldLeft(toLog) { case (unit, err) => unit.log(err) }
-      events.runEmptyA.written.unsafeRunSync shouldEqual errors.reverse
+      events.written.unsafeRunSync shouldEqual errors.reverse
     }
   }
 }
